First cut of simple template support.
The idea is that this allows a user to write a simple class which uses
mako to generate a function.  This should simplify auto-generating some
of the functions we require with a relatively simple syntax.  This
requires a small hack as the function we generate is dynamic and our
code generation needs to get the source code, we do this by setting a
`source` attribute on the generated function.