Fix a bug causing a PySPH regression.
This was caused by the recent JIT reorganization.  The issue is when a
jitted function calls another function which is annotated explicitly.
When this happens, the annotation helper was clobbering the explicit
type information and this led to incorrect wrapping of a function.  We
now mark jitted functions as such and for functions for which explicit
annotations are provided the jitting does not re-annotate the function
incorrectly.