ENH: Handle literal float values carefully.
When `use_double` is false, this represents literal floats as 32 bit
floats otherwise it leaves it as such. This is essential for good
performance on GPUs with limited or no fp64 capability. It also
enormously improves performance when doubles are not used on such GPUs.
This also eliminates the need for unnecessary casting in the examples to
attain good performance.